Add Unique constraint to Users, fix HasKey issue

+++ back/Cumpuss.Infrastructure/Models/User.cs
@@ -6,7 +6,7 @@
         public int FaceId { get; set; }
         public string Login { get; set; }
         public string PasswordHash { get; set; }
-        public short UserType { get; set; }
+        public int UserType { get; set; }
 
         public virtual Person Face { get; set; }
     }
